Working fsk (#51)
* Started FSK testing

* This is not working at all, saving to try another library

* Everything is WORKING, FSK is AWESOME!!

* FSK working with scan

* Revert "Merge branch '18-feat-implement-fsk' into working-fsk"

This reverts commit a1879c76d3937288460cbe56829ed19afea9f80d, reversing
changes made to eed6247545fa59fd1932932b3ff8d9aa1eb1d24a.

* Generated files

* Moved all functions out of main

* Moved heaaders to fsk file

* Moved externs

* Fix TIM15 initialization parameters

* Update PWM pulse value for TIM_OCMODE_PWM1

* Change PWM_HI calculation to 75% duty cycle

* Change TIM_NUM definition from 1 to 0

* Update gamma correction and timer number definitions

* Disable gamma correction and update timer number

* Refactor init_fsk and start_fsk functions

* Modify FSK functions to accept pwm_data

Updated init_fsk and start_fsk functions to include pwm_data parameter.

---------
