+++ Core/User/FSK.h
@@ -9,8 +9,8 @@
 extern TIM_HandleTypeDef htim2;
 extern TIM_HandleTypeDef htim15;
 
-void init_fsk(uint8_t* msg, uint32_t *pwm_data);
+void init_fsk(uint8_t* msg);
 
-void start_fsk(uint32_t *pwm_data);
+void start_fsk(void);
 
 void stop_fsk(void);
