--- Core/User/led.h
@@ -1,18 +1,0 @@
-#ifndef LED_H
-#define LED_H
-
-#include "math.h"
-#include "ARGB.h"   
-#include "stm32g0xx_hal.h"
-
-/* Easy definition of LEDs
-#define PORT 0
-#define BOW 1
-#define STARBOARD 2
-#define PORT 3
-#define DISTANCE 4
-*/
-
-void led_main(void);
-
-#endif
